For dev

+++ src/components/Main/BookBoard.js
@@ -1,6 +1,6 @@
 import React, {Component} from "react"
 import { Link } from 'react-router-dom'
-import Images, {Image} from 'react-bootstrap'
+import {Image} from 'react-bootstrap'
 import { Icon } from "semantic-ui-react";
 import server_url from '../../url.json'
 import axios from 'axios'
@@ -10,7 +10,7 @@ class BookBoard extends Component {
 	state = {
 			liked: false,
 			likeCount: this.props.likecount
-	}
+	} 
 
 	token = window.localStorage.getItem('token')
 
@@ -31,6 +31,11 @@ class BookBoard extends Component {
 
 
 	_handleLike = () => {
+		if(this.props.userName===''){
+			alert("유져네임을 설정해주세요")
+			return;
+		}
+
 		if (this.state.liked) {
 			axios.delete(`https://${server_url}/api/like/${this.props.postid}`, {
 				headers: {Authorization: `bearer ${this.token}`}
@@ -68,26 +73,31 @@ class BookBoard extends Component {
 			this._handleLike()
 	} */
 
+	handleImage =()=>{
+		if(this.props.url===''&&this.props.bookData!=='null'){
+			let parsedBookData = JSON.parse(this.props.bookData);
+			let postImage = parsedBookData.cover;
+			return postImage 
+		} 
+		return `https://${server_url}/upload/${this.props.url}`	
+	}	
+
 	render(){
 		return(
-				<div className ='BookBoard'>
-					<Link to={{
-						pathname : `/postdetail/${this.props.postid}`
-					}}>
-					<div className='postImageContainer'>
-						<Image className = 'mainThumbNail' alt='bookcover' /* width={240} height={240} */
-									src = {`https://${server_url}/upload/${this.props.url}`}/>
-									
-						<div className='likePart'>
-							<span className='likeIconPart'>{this.state.liked
+				<div className ='bookBoard'>
+					<div className='imageContainer'>
+						<Link to={{ pathname : `/postdetail/${this.props.postid}` }}>
+							<Image className = 'mainThumbNail' alt='bookcover' /* width={240} height={240} */
+										src = {this.handleImage()}/>
+						</Link>
+						<div className='likeBar'>
+							{this.state.liked
 							? <div><Icon name='heart' size="large" onClick={this._handleLike}/>{this.state.likeCount}</div>
 							: <div><Icon name='heart outline' size="large" onClick={this._handleLike}/>{this.state.likeCount}</div>
-							}</span>
+							}
 						</div>
 					</div>
-					</Link>
 					<p className='postTitle'>{this.props.title}</p>
-					<span className='userNamePart'>{this.props.username}</span>
 				</div>
 		)
 	}
