+++ src/components/PostDetail/PostWriter.js
@@ -1,17 +1,15 @@
 import React, { Component } from 'react';
-import { Button, button } from "semantic-ui-react";
+import { Button} from "semantic-ui-react";
 import axios from 'axios';
-import path from 'path';
 import server_url from '../../url.json';
-//import "./PostDetail.css";
-import { pathToFileURL } from 'url';
 import profileimage from "../../img/다운로드.png"
 import { PropTypes} from 'prop-types';
-
+import { Link} from "react-router-dom";
 
 export default class PostWriter extends Component {
 
   state = {
+    userId: '',
     userName: '',
     userImage: '',
     isFollowing:false,
@@ -27,13 +25,20 @@ export default class PostWriter extends Component {
   }
 
   _getUserData = async () => {
-    console.log("user 데이터를 가져와야하는데?")
     const res_getUser = await axios.get(`https://${server_url}/api/post/postedUserName/${this.props.postId}`, this.authHeader)
-    console.log('_getUserData에서 res_getUser =========',res_getUser)
+    console.log('res_getUser',res_getUser)
+    let userName = null
+    let profileImage =null
+    if(res_getUser.data[1]){
+      userName= res_getUser.data[1].userName
+      profileImage= res_getUser.data[1].profileImage
+    }
+    let userId= res_getUser.data[0]
+
     this.setState({
-      userName: res_getUser.data.userName,
-      userImage: res_getUser.data.profileImage,
-      userId: res_getUser.data.id
+      userName: userName,
+      userImage: profileImage,
+      userId: userId
     })
   }
 
@@ -44,23 +49,30 @@ export default class PostWriter extends Component {
   }
 
   _handleFollowing = async () => {
+
+    if(this.props.userName === ''){
+      alert("유저네임을 설정해주세요")
+      return;
+    }
+
+
     if (this.state.isFollowing) {
 
       const resultfollowDelete =await axios.delete(`https://${server_url}/api/follow/delete/${this.props.userId}`, this.authHeader)
-      console.log ("resultfollowDelete",resultfollowDelete)
+
       this.setState({isFollowing: false})
     } else {
       
       const followResult =await axios.post(`https://${server_url}/api/follow/${this.props.userId}`, {}, this.authHeader)
-      console.log("followResult", followResult)
+
       this.setState({isFollowing: true})
     }
   }
 
   _handleDelete = async () => {
     const res_deletePost = await axios.delete(`https://${server_url}/api/post/${this.props.postId}`, this.authHeader)
-    console.log("props", this.props)
-    console.log(res_deletePost.data,'삭제되었습니다. res_deletePost.data');
+    // console.log("props", this.props)
+    // console.log(res_deletePost.data,'삭제되었습니다. res_deletePost.data');
     window.location.href ="/mypage"
   }
 
@@ -69,36 +81,36 @@ export default class PostWriter extends Component {
   }
 
   _userImagecontrollor =()=>{
-    console.log("userImage는?",this.state.userImage)
-      if(this.state.userImage!==null && this.state.userImage.length>0){
-      return  <img src={`https://${server_url}/upload/${this.state.userImage}`} className='img-circle' alt={"user?"} />
+      if(this.state.userImage){
+      return <img src={`https://${server_url}/upload/${this.state.userImage}`} className='img-circle' alt={"user?"} />
       } else {
       return  <img src= {profileimage} className='img-circle' alt={"userImages"} />
       }  
     }
   
   render() {
-    const {userImage, userName, isFollowing} = this.state
+    const {userName, isFollowing, userId} = this.state
+    console.log("userId",userId)
     return (
       <div className='post_detail_right_1_postWriter'>
-        <div className="user_thumbs">{this._userImagecontrollor()}</div>
-        <h3 className='post_detail_username'>{userName}</h3>
-        <div className="user_buttons">
+        <Link to={`/postWriter/${userId}`}>
+          <div className="user_thumbs">{this._userImagecontrollor()}</div>
+          <div className='post_detail_username'>{userName}</div>
+        </Link>
         {(this.props.isMypost) //내 POST이면, 팔로우/팔로잉 을 보여주지 않고, post수정/삭제 를 보여줍니다.
         ? 
-          <div>
+          <div className="user_buttons">
             <Button inverted color='black' onClick={this._handleEdit}>수정</Button>
             <Button inverted color='grey' onClick={this._handleDelete}>삭제</Button>
           </div>
         : 
-          <div>
+          <div className="user_follow">
             {(isFollowing)
-            ? <button className="ui teal button" onClick={this._handleFollowing}>팔로우중입니다</button>
-            : <button className="ui grey button" onClick={this._handleFollowing}>팔로우하기</button> 
+            ? <button className="ui teal button" onClick={this._handleFollowing}>팔로잉</button>
+            : <button className="ui grey button" onClick={this._handleFollowing}>팔로우</button> 
             }
           </div>
         }
-        </div>
       </div>
     )
   }
